More cleanup, build using standard-issue OpenMP library.
Part 1 of multithreading dectris_stream to handle the likely 5-12GB/s of
incoming data it will receive from the upper end of large datasets.

Also setting the table for handing off data from bparchived to bpcompressd
and bpindexd. Whether each should receive images by reading the output or
the in-memory copy of the data depends on the feasibility of running bparchived,
bpindexd, and bpcompressd on the same machine. This will most likely be
impossible considering we do not have a single machine with all the necessary
compute power to programatically identify and index the very best two images
in a massive dataset in real-time.

All we can do is try to make the API as modular as possible right now in the
infancy (2nd month) of development.